server/requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
Gunicorn
tensorflow==2.11.0
opencv-python==4.8.0.76
numpy==1.21.6
Pillow==9.2.0
Werkzeug==2.3.6